+++ ext_tables.sql
@@ -3,6 +3,7 @@ CREATE TABLE tx_filetransfer_domain_model_upload (
 	receiver_address VARCHAR(255) DEFAULT '' NOT NULL,
 	subject VARCHAR(255) DEFAULT '' NOT NULL,
 	message MEDIUMTEXT,
+	signature MEDIUMTEXT,
 	token VARCHAR(80) DEFAULT '' NOT NULL,
 	download_limit INT(11) DEFAULT '1' NOT NULL,
 	asset INT(10) DEFAULT '0' NOT NULL,